+++ wax_uninstall.sh
@@ -16,7 +16,7 @@ if [ -d "/usr/local/eosio" ]; then
       case $yn in
          [Yy]* )
             if [ "$(id -u)" -ne 0 ]; then
-               printf "\n\tThis requires sudo, please run ./eosio_uninstall.sh with sudo\n\n"
+               printf "\n\tThis requires sudo, please run ./wax_uninstall.sh with sudo\n\n"
                exit -1
             fi