+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pending (wax-1.6.1-X.Y.Z)
 
 BREAKING CHANGES:
+- [KEW-1316] Renaming EOSIO with WAX.
 
 FEATURES: